refactor: change same name modules
