--- game_logic/player_commands.py
@@ -1,3 +1,0 @@
-class PlayerCommands:
-    def move_snake(self, snake_id, direction):
-        pass
